requirements.txt
requests
pandas
python-whois
dnspython
